+++ test/integration/embedders/test_octoai.py
@@ -4,9 +4,17 @@ from pathlib import Path
 
 import pytest
 
-from test.integration.embedders.utils import validate_embedding_output, validate_raw_embedder
+from test.integration.embedders.utils import (
+    validate_embedding_output,
+    validate_raw_embedder,
+    validate_raw_embedder_async,
+)
 from test.integration.utils import requires_env
-from unstructured_ingest.embed.octoai import OctoAiEmbeddingConfig, OctoAIEmbeddingEncoder
+from unstructured_ingest.embed.octoai import (
+    AsyncOctoAIEmbeddingEncoder,
+    OctoAiEmbeddingConfig,
+    OctoAIEmbeddingEncoder,
+)
 from unstructured_ingest.v2.errors import UserAuthError
 from unstructured_ingest.v2.processes.embedder import Embedder, EmbedderConfig
 
@@ -53,3 +61,17 @@ def test_raw_octoai_embedder_invalid_credentials():
     )
     with pytest.raises(UserAuthError):
         embedder.get_exemplary_embedding()
+
+
+@requires_env(API_KEY)
+@pytest.mark.asyncio
+async def test_raw_async_octoai_embedder(embedder_file: Path):
+    api_key = get_api_key()
+    embedder = AsyncOctoAIEmbeddingEncoder(
+        config=OctoAiEmbeddingConfig(
+            api_key=api_key,
+        )
+    )
+    await validate_raw_embedder_async(
+        embedder=embedder, embedder_file=embedder_file, expected_dimensions=(1024,)
+    )