+++ test/integration/embedders/test_vertexai.py
@@ -2,9 +2,19 @@ import json
 import os
 from pathlib import Path
 
-from test.integration.embedders.utils import validate_embedding_output, validate_raw_embedder
+import pytest
+
+from test.integration.embedders.utils import (
+    validate_embedding_output,
+    validate_raw_embedder,
+    validate_raw_embedder_async,
+)
 from test.integration.utils import requires_env
-from unstructured_ingest.embed.vertexai import VertexAIEmbeddingConfig, VertexAIEmbeddingEncoder
+from unstructured_ingest.embed.vertexai import (
+    AsyncVertexAIEmbeddingEncoder,
+    VertexAIEmbeddingConfig,
+    VertexAIEmbeddingEncoder,
+)
 from unstructured_ingest.v2.processes.embedder import Embedder, EmbedderConfig
 
 API_KEY = "VERTEXAI_API_KEY"
@@ -37,5 +47,19 @@ def test_raw_vertexai_embedder(embedder_file: Path):
         )
     )
     validate_raw_embedder(
+        embedder=embedder, embedder_file=embedder_file, expected_dimensions=(768,)
+    )
+
+
+@requires_env(API_KEY)
+@pytest.mark.asyncio
+async def test_raw_async_vertexai_embedder(embedder_file: Path):
+    api_key = get_api_key()
+    embedder = AsyncVertexAIEmbeddingEncoder(
+        config=VertexAIEmbeddingConfig(
+            api_key=api_key,
+        )
+    )
+    await validate_raw_embedder_async(
         embedder=embedder, embedder_file=embedder_file, expected_dimensions=(768,)
     )