+++ unstructured_ingest/embed/vertexai.py
@@ -8,8 +8,13 @@ from typing import TYPE_CHECKING, Annotated, Any, Optional
 from pydantic import Field, Secret, ValidationError
 from pydantic.functional_validators import BeforeValidator
 
-from unstructured_ingest.embed.interfaces import BaseEmbeddingEncoder, EmbeddingConfig
+from unstructured_ingest.embed.interfaces import (
+    AsyncBaseEmbeddingEncoder,
+    BaseEmbeddingEncoder,
+    EmbeddingConfig,
+)
 from unstructured_ingest.utils.dep_check import requires_dependencies
+from unstructured_ingest.v2.errors import UserAuthError
 
 if TYPE_CHECKING:
     from vertexai.language_models import TextEmbeddingModel
@@ -31,6 +36,13 @@ class VertexAIEmbeddingConfig(EmbeddingConfig):
     embedder_model_name: Optional[str] = Field(
         default="textembedding-gecko@001", alias="model_name"
     )
+
+    def wrap_error(self, e: Exception) -> Exception:
+        from google.auth.exceptions import GoogleAuthError
+
+        if isinstance(e, GoogleAuthError):
+            return UserAuthError(e)
+        return e
 
     def register_application_credentials(self):
         # TODO look into passing credentials in directly, rather than via env var and tmp file
@@ -55,6 +67,9 @@ class VertexAIEmbeddingConfig(EmbeddingConfig):
 class VertexAIEmbeddingEncoder(BaseEmbeddingEncoder):
     config: VertexAIEmbeddingConfig
 
+    def wrap_error(self, e: Exception) -> Exception:
+        return self.config.wrap_error(e=e)
+
     def embed_query(self, query):
         return self._embed_documents(elements=[query])[0]
 
@@ -70,7 +85,42 @@ class VertexAIEmbeddingEncoder(BaseEmbeddingEncoder):
     def _embed_documents(self, elements: list[str]) -> list[list[float]]:
         from vertexai.language_models import TextEmbeddingInput
 
-        client = self.config.get_client()
         inputs = [TextEmbeddingInput(text=element) for element in elements]
-        embeddings = client.get_embeddings(inputs)
+        try:
+            client = self.config.get_client()
+            embeddings = client.get_embeddings(inputs)
+        except Exception as e:
+            raise self.wrap_error(e=e)
+        return [e.values for e in embeddings]
+
+
+@dataclass
+class AsyncVertexAIEmbeddingEncoder(AsyncBaseEmbeddingEncoder):
+    config: VertexAIEmbeddingConfig
+
+    def wrap_error(self, e: Exception) -> Exception:
+        return self.config.wrap_error(e=e)
+
+    async def embed_query(self, query):
+        embedding = await self._embed_documents(elements=[query])
+        return embedding[0]
+
+    async def embed_documents(self, elements: list[dict]) -> list[dict]:
+        embeddings = await self._embed_documents([e.get("text", "") for e in elements])
+        elements_with_embeddings = self._add_embeddings_to_elements(elements, embeddings)
+        return elements_with_embeddings
+
+    @requires_dependencies(
+        ["vertexai"],
+        extras="embed-vertexai",
+    )
+    async def _embed_documents(self, elements: list[str]) -> list[list[float]]:
+        from vertexai.language_models import TextEmbeddingInput
+
+        inputs = [TextEmbeddingInput(text=element) for element in elements]
+        try:
+            client = self.config.get_client()
+            embeddings = await client.get_embeddings_async(inputs)
+        except Exception as e:
+            raise self.wrap_error(e=e)
         return [e.values for e in embeddings]