+++ CHANGELOG.md
@@ -1,16 +1,46 @@
-## 0.3.13-dev3
+## 0.3.16-dev1
 
 ### Enchancements
 
 * **Sharepoint coonner no longer uses depracted authorisation method. Uses EntraID instead via msal.**
 
-## 0.3.13-dev2
+## 0.3.16-dev0
+
+### Fixes
+
+* **Fix SQL uploader stager** - When passed `output_filename` without a suffix it resulted in unsupported file format error. Now, it will take a suffix of `elements_filepath` and append it to `output_filename`.
+* **Fix Snowflake uploader** - Unexpected `columns` argument was passed to `_fit_to_schema` method inside SnowflakeUploader `upload_dataframe` method.
+
+## 0.3.15
+
+### Enhancements
+
+* **Add databricks delta table connector**
+
+### Fixes
+
+* **Fixed namespace issue with pinecone, and added new test**
+
+## 0.3.14
+
+### Fixes
+
+* **Fix Neo4j Uploader string enum error**
+* **Fix ChromaDB Destination failing integration tests** - issue lies within the newest ChromaDB release, fix freezes it's version to 0.6.2.
+
+## 0.3.13
 
 ### Fixes
 
 * **Fix Snowflake Uploader error**
 * **Fix SQL Uploader Stager timestamp error**
 * **Migrate Discord Sourced Connector to v2**
+* **Add read data fallback** When reading data that could be json or ndjson, if extension is missing, fallback to trying to read it as json.
+
+### Enhancements
+
+* **Async support for all IO-bounded embedders**
+* **Expand support to Python 3.13**
 
 ## 0.3.12